+++ template/package.json
@@ -10,6 +10,7 @@
     "lint": "eslint ."
   },
   "dependencies": {
+    "@react-navigation/native": "^5.7.3",
     "react": "16.13.1",
     "react-native": "0.63.2"
   },
